Add tests of SqlAsyncWorker
